chore: move microphone-stream.js to src folder
