+++ src/microphone-stream.js
